#584: Don't reset viewport on selection mode change
