+++ src/netxs/desktopio/canvas.hpp
@@ -480,6 +480,24 @@ namespace netxs
                        + std::to_string(chan.b) + ","
                        + std::to_string(chan.a) + "}";
         }
+        // rgba: Linear to sRGB (g = 2.4)
+        static auto gamma(fp32 c)
+        {
+            return c <= 0.0031308f ? 12.92f * c
+                                   : 1.055f * std::pow(c, 1.f / 2.4f) - 0.055f;
+        }
+        // rgba: Linear to sRGB (g = 2.4)
+        static auto gamma(byte c)
+        {
+            return (byte)rgba::gamma(c / 255.f);
+        }
+        // rgba: Linear to sRGB (g = 2.4)
+        static auto gamma(rgba c)
+        {
+            return rgba{ rgba::gamma(c.chan.r),
+                         rgba::gamma(c.chan.g),
+                         rgba::gamma(c.chan.g), c.chan.a };
+        }
 
         template<si32 i>
         static constexpr ui32 _vt16 = // Compile-time value assigning (sorted by enum).
@@ -2120,6 +2138,231 @@ namespace netxs
                              : form == ansitext ? tag::ansi
                              : form == safetext ? tag::safe
                                                 : tag::text, "/", size.x, "/", size.y);
+        }
+    }
+
+    namespace misc //todo classify
+    {
+        template<class T, auto fx>
+        struct shadow
+        {
+            T  bitmap{};
+            bool sync{};
+            twod over{};
+            twod step{};
+
+            // shadow: Generate sRGB (gamma=2.2) shadow sprite.
+            void generate(fp32 bias, fp32 alfa, si32 size, twod offset, twod ratio, auto fuse)
+            {
+                //bias    += _k0 * 0.1f;
+                //opacity += _k1 * 1.f;
+                //size    += _k2;
+                //offset  +=  ratio * _k3;
+                sync = true;
+                alfa = std::clamp(alfa, 0.f, 255.f);
+                size = std::abs(size) * 2;
+                over = ratio * (size * 2);
+                step = over / 2 - offset;
+                auto spline = netxs::spline01{ bias };
+                auto sz = ratio * (size * 2 + 1);
+                bitmap.size(sz);
+                auto it = bitmap.begin();
+                for (auto y = 0.f; y < sz.y; y++)
+                {
+                    auto y0 = y / (sz.y - 1.f);
+                    auto sy = spline(y0);
+                    for (auto x = 0.f; x < sz.x; x++)
+                    {
+                        auto x0 = x / (sz.x - 1.f);
+                        auto sx = spline(x0);
+                        auto xy = sy * sx; // rgba::gamma(sy * sx);
+                        auto a = (byte)std::round(alfa * xy);
+                        fuse(*it++, a);
+                    }
+                }
+            }
+            auto render(auto& canvas, auto win_size)
+            {
+                canvas.step(step);
+                auto src = bitmap.area();
+                auto dst = rect{ dot_00, win_size + over };
+                auto cut = std::min(dot_00, (dst.size - src.size * 2 - dot_11) / 2);
+                auto off = dent{ 0, cut.x, 0, cut.y };
+                src += off;
+                auto mid = rect{ src.size, std::max(dot_00, dst.size - src.size * 2) };
+                auto top = rect{ twod{ src.size.x, 0 }, { mid.size.x, src.size.y }};
+                auto lft = rect{ twod{ 0, src.size.y }, { src.size.x, mid.size.y }};
+                if (mid)
+                {
+                    auto base_shadow = bitmap[src.size - dot_11];
+                    netxs::onrect(canvas, mid, fx(base_shadow));
+                }
+                if (top)
+                {
+                    auto pen = rect{{ src.size.x - 1, 0 }, { 1, src.size.y }};
+                    netxs::xform_scale(canvas, top, bitmap, pen, fx);
+                    top.coor.y += mid.size.y + top.size.y;
+                    netxs::xform_scale(canvas, top, bitmap, pen.rotate({ 1, -1 }), fx);
+                }
+                if (lft)
+                {
+                    auto pen = rect{{ 0, src.size.y - 1 }, { src.size.x, 1 }};
+                    netxs::xform_scale(canvas, lft, bitmap, pen, fx);
+                    lft.coor.x += mid.size.x + lft.size.x;
+                    netxs::xform_scale(canvas, lft, bitmap, pen.rotate({ -1, 1 }), fx);
+                }
+                auto dir = dot_11;
+                            netxs::xform_mirror(canvas, dst.rotate(dir).coor, bitmap, src.rotate(dir), fx);
+                dir = -dir; netxs::xform_mirror(canvas, dst.rotate(dir).coor, bitmap, src.rotate(dir), fx);
+                dir.x += 2; netxs::xform_mirror(canvas, dst.rotate(dir).coor, bitmap, src.rotate(dir), fx);
+                dir = -dir; netxs::xform_mirror(canvas, dst.rotate(dir).coor, bitmap, src.rotate(dir), fx);
+                canvas.step(-step);
+            }
+        };
+        struct szgrips
+        {
+            using test = testy<twod>;
+
+            twod origin; // szgrips: Grab's initial coord info.
+            twod dtcoor; // szgrips: The form coor parameter change factor while resizing.
+            twod sector; // szgrips: Active quadrant, x,y = {-1|+1}. Border widths.
+            rect hzgrip; // szgrips: Horizontal grip.
+            rect vtgrip; // szgrips: Vertical grip.
+            twod widths; // szgrips: Grip's widths.
+            bool inside; // szgrips: Is active.
+            bool seized; // szgrips: Is seized.
+            rect zoomsz; // szgrips: Captured area for zooming.
+            dent zoomdt; // szgrips: Zoom step.
+            bool zoomon; // szgrips: Zoom in progress.
+            twod zoomat; // szgrips: Zoom pivot.
+
+            szgrips()
+                : inside{ faux },
+                  seized{ faux },
+                  zoomon{ faux }
+            { }
+
+            operator bool () { return inside || seized; }
+            auto corner(twod length) const
+            {
+                return dtcoor.less(dot_11, length, dot_00);
+            }
+            auto quantize(twod curpos, twod basis, twod cell_size) const
+            {
+                curpos -= basis;
+                curpos -= (curpos + cell_size/*to avoid negative values*/) % cell_size;
+                return curpos;
+            }
+            auto grab(rect window, twod curpos, dent outer, twod cell_size = dot_11)
+            {
+                if (inside)
+                {
+                    auto outer_rect = window + outer;
+                    curpos = quantize(curpos, outer_rect.coor, cell_size);
+                    origin = curpos - corner(outer_rect.size);
+                    seized = true;
+                }
+                return seized;
+            }
+            auto calc(rect window, twod curpos, dent outer, dent inner, twod cell_size = dot_11)
+            {
+                auto border = outer - inner;
+                auto inner_rect = window + inner;
+                auto outer_rect = window + outer;
+
+                auto inside_prev = inside;
+                auto hzgrip_prev = hzgrip;
+                auto vtgrip_prev = vtgrip;
+
+                inside = !inner_rect.hittest(curpos)
+                       && outer_rect.hittest(curpos);
+                if (!seized)
+                {
+                    auto r1 = inner_rect;
+                    auto l1 = r1.coor - curpos;
+                    auto l2 = r1.coor + r1.size - curpos;
+                    auto m = std::min({ std::abs(l1.x), std::abs(l1.y), std::abs(l2.x), std::abs(l2.y) });
+                    m == std::abs(l1.x) ? curpos.x = r1.coor.x :
+                    m == std::abs(l1.y) ? curpos.y = r1.coor.y :
+                    m == std::abs(l2.x) ? curpos.x = r1.coor.x + r1.size.x :
+                                          curpos.y = r1.coor.y + r1.size.y;
+                }
+                auto& length = outer_rect.size;
+                curpos = quantize(curpos, outer_rect.coor, cell_size);
+                auto center = std::max(length / 2, dot_11);
+                if (!seized)
+                {
+                    dtcoor = curpos.less(center + (length & 1), dot_11, dot_00);
+                    sector = dtcoor.less(dot_11, -dot_11, dot_11);
+                    widths = sector.less(dot_00, twod{-border.r,-border.b },
+                                                 twod{ border.l, border.t });
+                }
+                auto l = sector * (curpos - corner(length));
+                auto a = center * l / center;
+                auto b = center *~l /~center;
+                auto s = sector * std::max(a - b + center, dot_00);
+
+                hzgrip.coor.x = widths.x;
+                hzgrip.coor.y = 0;
+                hzgrip.size.y = widths.y;
+                hzgrip.size.x = s.x - s.x % cell_size.x;
+
+                vtgrip.coor = dot_00;
+                vtgrip.size = widths;
+                vtgrip.size.y += s.y - s.y % cell_size.y;
+                auto changed = inside_prev != inside || (inside && (hzgrip_prev != hzgrip || vtgrip_prev != vtgrip));
+                return changed;
+            }
+            auto drag(rect window, twod curpos, dent outer, bool zoom, twod cell_size = dot_11) const
+            {
+                auto outer_rect = window + outer;
+                curpos = quantize(curpos, outer_rect.coor, cell_size);
+                auto delta = (corner(outer_rect.size) + origin - curpos) * sector;
+                if (zoom) delta *= 2;
+                auto preview_step = zoom ? -delta / 2 : -delta * dtcoor;
+                auto preview_area = rect{ window.coor + preview_step, window.size + delta };
+                return std::pair{ preview_area, delta };
+            }
+            auto move(twod dxdy, bool zoom) const
+            {
+                auto step = zoom ? -dxdy / 2 : -dxdy * dtcoor;
+                return step;
+            }
+            void drop()
+            {
+                seized = faux;
+            }
+            auto draw(auto& canvas, rect area, auto fx) const
+            {
+                auto vertex = corner(area.size);
+                auto side_x = hzgrip.shift(vertex).normalize_itself().shift_itself(area.coor).trim(area);
+                auto side_y = vtgrip.shift(vertex).normalize_itself().shift_itself(area.coor).trim(area);
+                netxs::onrect(canvas, side_x, fx);
+                netxs::onrect(canvas, side_y, fx);
+                return std::pair{ side_x, side_y };
+            }
+        };
+
+        void fill(auto& canvas, auto block, auto fx) // gfx: Fill block.
+        {
+            block.normalize_itself();
+            netxs::onrect(canvas, block, fx);
+        }
+        void cage(auto& canvas, rect area, dent border, auto fx) // core: Draw the cage around specified area.
+        {
+            auto temp = area;
+            temp.size.y = std::max(0, border.t); // Top
+            fill(canvas, temp.trim(area), fx);
+            temp.coor.y += area.size.y - border.b; // Bottom
+            temp.size.y = std::max(0, border.b);
+            fill(canvas, temp.trim(area), fx);
+            temp.size.x = std::max(0, border.l); // Left
+            temp.size.y = std::max(0, area.size.y - border.t - border.b);
+            temp.coor.y = area.coor.y + border.t;
+            fill(canvas, temp.trim(area), fx);
+            temp.coor.x += area.size.x - border.r; // Right
+            temp.size.x = std::max(0, border.r);
+            fill(canvas, temp.trim(area), fx);
         }
     }
